+++ CHANGELOG.md
@@ -1,4 +1,21 @@
+## 2.1.0 (2022-04-17)
+
+### Features
+
+- Modify API:
+
+  | Old API              | New API         |
+  | -------------------- | --------------- |
+  | `strveRouterVersion` | `routerVersion` |
+  | `StrveRouter`        | `initRouter`    |
+  | `param2Obj`          | `toParse`       |
+  | `routerLink`         | `linkTo`        |
+
+- Remove API `routerHashUpdate`;
+- TypeScript refactoring code;
+
 ## 2.0.4 (2022-02-26)
+
 ### Features
 
 - Update `routerLink` API;